// src/app/shared/models/constants.ts
export enum SengaType{
  AUDIO = 'audio',
  VIDEO = 'video',
  PHOTO = 'photo',
  TEXT = 'text'
}

export enum SengaTypeIcons{
  AUDIO = 'microphone',
  VIDEO = 'video-camera',
  PHOTO ='camera',
  TEXT = 'pencil-square-o'
}

export enum SengaTypeColors{
  AUDIO = 'f7437f',
  VIDEO = 'b330f0',
  PHOTO ='30f0c0',
  TEXT = '20a8f7'
}